requirements.txt
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
reportlab>=4.0.0
flask>=3.0.0
python-magic-bin>=0.4.14; sys_platform == "win32"
python-magic>=0.4.27; sys_platform != "win32"
